# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "D-Care Console API",
    "description": "치과 아웃바운드 관리 시스템 API\n\n## 인증\n모든 V2 API는 JWT Bearer 토큰 인증이 필요합니다.\n`Authorization: Bearer <token>` 헤더를 포함하세요.",
    "version": "2.0.0"
  },
  "servers": [
    { "url": "/", "description": "현재 서버" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Patient": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "phone": { "type": "string" },
          "status": { "type": "string", "enum": ["consulting", "reserved", "visited", "treatmentBooked", "treatment", "completed", "followup", "closed"] },
          "temperature": { "type": "string", "enum": ["hot", "warm", "cold"] },
          "interest": { "type": "string" },
          "source": { "type": "string" },
          "consultationType": { "type": "string" },
          "age": { "type": "integer" },
          "urgency": { "type": "string", "enum": ["noshow", "today", "overdue", "normal"] },
          "daysInStatus": { "type": "integer" },
          "nextAction": { "type": "string" },
          "nextActionDate": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "lastContactAt": { "type": "string", "format": "date-time" }
        }
      },
      "CallLog": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "phone": { "type": "string" },
          "callType": { "type": "string", "enum": ["inbound", "outbound"] },
          "duration": { "type": "integer" },
          "callTime": { "type": "string", "format": "date-time" },
          "patientId": { "type": "string" },
          "patientName": { "type": "string" },
          "classification": { "type": "string", "enum": ["환자", "거래처", "스팸", "기타"] },
          "summary": { "type": "string" },
          "interest": { "type": "string" },
          "temperature": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "processing", "completed"] },
          "callbackType": { "type": "string", "enum": ["callback", "recall", "thanks"] }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v2/patients/{id}": {
      "get": {
        "tags": ["Patients"],
        "summary": "환자 상세 조회",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "환자 정보 + 최근 통화 10건",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "patient": { "$ref": "#/components/schemas/Patient" },
                    "callLogs": { "type": "array", "items": { "$ref": "#/components/schemas/CallLog" } }
                  }
                }
              }
            }
          },
          "404": { "description": "환자를 찾을 수 없음" }
        }
      },
      "patch": {
        "tags": ["Patients"],
        "summary": "환자 정보 수정",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "phone": { "type": "string" },
                  "status": { "type": "string" },
                  "temperature": { "type": "string" },
                  "interest": { "type": "string" },
                  "nextAction": { "type": "string" },
                  "nextActionDate": { "type": "string", "format": "date-time" },
                  "memo": { "type": "string" },
                  "estimatedAmount": { "type": "number" },
                  "actualAmount": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "수정 성공", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "404": { "description": "환자를 찾을 수 없음" }
        }
      },
      "delete": {
        "tags": ["Patients"],
        "summary": "환자 삭제",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "삭제 성공" },
          "404": { "description": "환자를 찾을 수 없음" }
        }
      }
    },
...
  },
...
}